Fixes an issue with bool calculation, and adds the non-indirection version of sum

+++ include/qdp_precision.h
@@ -19,7 +19,7 @@
 // These are fixed precision versions
 typedef int       INTEGER32;
 typedef float     REAL32;
-typedef double    REAL64;
+typedef float    REAL64;   // double
 typedef bool      LOGICAL;
 
 // Set the base floating precision
